Refactor CLI modules and clean repository
