--- .git-scanner/history.json
@@ -1,8 +1,0 @@
-[
-  {
-    "path": "/Users/liorkirshner/Documents/TechProjects/git-scanner",
-    "date": "2025-07-28T09:40:49.722Z",
-    "status": "🟡",
-    "changes": []
-  }
-]
