+++ scanner.js
@@ -1,12 +1,10 @@
 #!/usr/bin/env node
 
-const path = require("path");
-const fs = require("fs");
 const { execSync } = require("child_process");
-const { ask, handleRepo, rl, close } = require("./utils/ask.js");
+const { ask, rl, close } = require("./utils/ask.js");
+const { handleRepo } = require("./utils/repoActions");
 const findGitRepos = require("./utils/gitScanner");
-const { updateHistoryEntry } = require("./utils/historyHandler");
-const { loadHistory } = require("./utils/historyHandler");
+const { updateHistoryEntry, loadHistory } = require("./utils/historyHandler");
 
 async function main() {
   console.clear();
@@ -93,7 +91,6 @@ async function main() {
 
     if (isNaN(index) || !reposWithStatus[index]) {
       console.log("👋 Bye.");
-      const { close } = require("./utils/ask.js");
       close();
       console.log("\n💡 You may now continue using the terminal.\n");
       break;
@@ -110,69 +107,6 @@ async function main() {
       console.log(result);
     } catch (e) {
       console.log("❌ Failed to run git status");
-    }
-
-    // For repos that are not green, display numbered options for user actions
-    const nonGreenRepos = reposWithStatus.filter(
-      (r) => r.statusSymbol === "🟡" || r.statusSymbol === "🔴"
-    );
-
-    for (let i = 0; i < nonGreenRepos.length; i++) {
-      const { repo, statusSymbol } = nonGreenRepos[i];
-      if (repo !== selected) continue;
-      console.log(`\nOptions for repo: ${repo}`);
-      console.log("Choose action:");
-      console.log("[1] Push changes (add + commit + push)");
-      console.log("[2] Add all changes (git add .)");
-      console.log("[3] Restore changes (git restore .)");
-      console.log("[4] Commit (only if already added)");
-      console.log("\x1b[31m[Enter] Cancel\x1b[0m"); // Red colored cancel
-      const action = await ask("> ");
-
-      switch (action.trim()) {
-        case "1":
-          const commitMsg = await ask("📝 Enter commit message: ");
-          try {
-            execSync(`git add . && git commit -m "${commitMsg}" && git push`, {
-              cwd: repo,
-              stdio: "inherit",
-            });
-            console.log("✅ Changes pushed successfully.");
-          } catch (e) {
-            console.log("❌ Failed to push changes.");
-          }
-          break;
-        case "2":
-          try {
-            execSync(`git add .`, { cwd: repo, stdio: "inherit" });
-            console.log("✅ Changes added.");
-          } catch (e) {
-            console.log("❌ Failed to add changes.");
-          }
-          break;
-        case "3":
-          try {
-            execSync(`git restore .`, { cwd: repo, stdio: "inherit" });
-            console.log("🔄 Changes restored.");
-          } catch (e) {
-            console.log("❌ Failed to restore.");
-          }
-          break;
-        case "4":
-          const msg = await ask("📝 Enter commit message: ");
-          try {
-            execSync(`git commit -m "${msg}"`, {
-              cwd: repo,
-              stdio: "inherit",
-            });
-            console.log("✅ Committed.");
-          } catch (e) {
-            console.log("❌ Commit failed.");
-          }
-          break;
-        default:
-          console.log("❌ Cancelled.");
-      }
     }
 
     await handleRepo(selected);
